+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed issue of finding parent class docstring and signature when not overridden by child class
 - Fixed module resolution for `classmethod`s and `staticmethod`s
 - Refactored `PythonAnalyzer` to cache file reads
+- Fixed parameter placement issue when parameter's were above `_target_`
 
 ## [0.1.4]